pyproject.toml
[project]
name = "xai-crewai"
version = "0.1.0"
description = "A Xircuits component library for CrewAI"
authors = [{ name = "XpressAI", email = "eduardo@xpress.ai" }]
license = "Apache-2.0"
readme = "README.md"
repository = "https://github.com/XpressAI/xai-crewai"
keywords = ["xircuits", "crewai", "agent", "AI"]

dependencies = [
    "crewai",
    "langchain-community",
    "Pillow",
    "langchain_openai",
    "python-dotenv",
]

# Xircuits-specific configurations
[tool.xircuits]
default_example_path = "examples/CrewAIBasicAgent.xircuits"
requirements_path = "requirements.txt"
